change routing from surveys to relative and rename manager dashboard module. Issue #6538
