+++ src/app/manager-dashboard/manager-dashboard.component.html
@@ -11,7 +11,7 @@
   <a *ngIf="planetType !== 'community' || isHub" [routerLink]="['reports', isHub ? { 'hubId': planetConfiguration._id } : {}]" i18n mat-raised-button>Reports</a>
   <a [routerLink]="['reports', 'myplanet', isHub ? { 'hubId': planetConfiguration._id } : {}]" i18n mat-raised-button>myPlanet Reports</a>
   <ng-container *planetAuthorizedRoles="'manager'">
-    <a [routerLink]="['/surveys']" mat-raised-button i18n>Surveys</a>
+    <a [routerLink]="['./surveys']" mat-raised-button i18n>Surveys</a>
     <button *ngIf="planetType !== center && showResendConfiguration"
       (click)="resendConfig()" i18n mat-raised-button>Resend Registration Request</button>
     <button *ngIf="devMode"
